Add files via upload
this project is mainly about borrowing books from the people who have read their books and don't want them anymore or didn't like them but it can be used to sell and buy books too. 
I have learned how to build a sign in/signout page along with the basics of making an android app, we also used java in this project, i also learned the importance of design alongside the functions used.